requirements/backend.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile backend.in
#
aiocsv==1.3.2
    # via -r backend.in
aiofiles==23.2.1
    # via
    #   -r backend.in
    #   aiogram
aiogram==3.6.0
    # via -r backend.in
aiohttp==3.9.5
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
apscheduler==3.10.4
    # via -r backend.in
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via -r backend.in
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   aiogram
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.110.3
    # via -r backend.in
fonttools==4.51.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via -r backend.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r backend.in
idna==3.7
    # via
    #   anyio
    #   httpx
    #   yarl
joblib==1.4.2
    # via
    #   -r backend.in
    #   scikit-learn
kiwisolver==1.4.5
    # via matplotlib
magic-filter==1.0.12
    # via aiogram
matplotlib==3.8.4
    # via -r backend.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   xgboost
packaging==24.0
    # via matplotlib
pandas==2.2.2
    # via -r backend.in
pillow==10.3.0
    # via matplotlib
psycopg2-binary==2.9.9
    # via -r backend.in
pydantic==2.7.1
    # via
    #   -r backend.in
    #   aiogram
    #   fastapi
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r backend.in
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.1
    # via
    #   apscheduler
    #   pandas
redis==5.0.4
    # via -r backend.in
schedule==1.2.1
    # via -r backend.in
scikit-learn==1.4.2
    # via -r backend.in
scipy==1.13.0
    # via
    #   scikit-learn
    #   xgboost
six==1.16.0
    # via
    #   apscheduler
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via -r backend.in
starlette==0.37.2
    # via fastapi
threadpoolctl==3.5.0
    # via scikit-learn
typing-extensions==4.11.0
    # via
    #   aiocsv
    #   aiogram
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via pandas
tzlocal==5.2
    # via apscheduler
uvicorn==0.29.0
    # via -r backend.in
xgboost==2.0.3
    # via -r backend.in
yarl==1.9.4
    # via aiohttp
